upgrade detik-prefect bisa append/concate

+++ prefect/flows/detik-prefect-test.py
@@ -48,46 +48,47 @@ def transform_data(df):
 
     column_names = ["generated_id", "name", "link_url", "author", "waktu_post", "category"]
     df.columns = column_names
-    print(f"columns :{df.dtypes}")
+    # print(f"columns :{df.dtypes}")
     return df
 
 @task(name="Write Local", log_prints=True)
 def write_local(df: pd.DataFrame, dataset_file: str):
+
+
     path = Path(f"data/{dataset_file}.parquet")
-    df.to_parquet(path, compression="gzip")
+    # print(path)
+    if path.is_file() == True:
+        df_before = pd.read_parquet(f"{path}")
+        df_new = pd.concat([df_before, df])
+        df_new.to_parquet(path, compression="gzip")
+        print(df_new)
+        print('hello')
+    else:
+        df.to_parquet(path, compression="gzip")
+        print('world')
+    
     return path
 
-@task()
-def write_gcs(path: Path) -> None:
-    """upload parquet to GCS"""
-    gcs_block = GcsBucket.load("bucket-google")
-    gcs_block.upload_from_path(
-        from_path=f"{path}",
-        to_path=path
-    )
-    """remove existing data on file system"""
-    cmd =F"rm {path}"
-    os.system(cmd)
 
 @flow(name="Scrap Flow", log_prints=True)
 def main_flow():
 
+    """flow scrap"""
     headers = {
         'user-agent': 'Mozilla/5.0 (Linux; Android 6.0; Nexus 5 Build/MRA58N) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/109.0.0.0 Mobile Safari/537.36'}
-
     key = 'pemilu+2024'
     now = datetime.now()
     date = now.strftime("%d/%m/%Y")  # dd/mm/yyyy
-
     url = f"https://www.detik.com/search/searchall?query={key}&sortby=time&sorttime=0&fromdatex={date}&todatex={date}&siteid=2"
+
 
     raw_data = scrap_task(url, key, headers)
     df = transform_data(raw_data)
 
-    dataset_date = now.strftime("%d_%m_%Y_%H_%M_%S")
+    dataset_date = now.strftime("%d_%m_%Y")
     dataset_file = f"detik_{dataset_date}"
     path = write_local(df, dataset_file)
-    write_gcs(path)
+    # write_gcs(path)
 
 
 if __name__ == '__main__':
